requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
google-generativeai==0.4.0
atlassian-python-api==3.41.11
slack-sdk==3.27.1
pandas==2.2.0
python-dotenv==1.0.1
tabulate==0.9.0
